Move bigint tests to filecheck
